// src/app/api/ui/enums/filter-application-state-alt.enum.ts
export enum FilterApplicationStateAltEnum {
  Formation = 1,
  Complete = 2,
  Rejected = 3,
  Approved = 4,
  Accepted = 5,
  RejectedByPart1 = 6,
  AcceptedByPart1 = 7,
  NotImplemented = 8,
  NotConsideredPart2 = 9,
  RejectedByPart2 = 10,
  AcceptedByPart2 = 11,
  Loser = 12,
  Winner = 13,
  ContractTerminated = 14,
  Withdrawn = 15,
  RejectedTab = 16,
  NotActual = 17,
  WaitingForFinalProposal = 18,
  FinalProposalSubmitted = 19,
}